requirements.txt
openpyxl>=3.1.2
typer>=0.12.0
rich>=13.7.0
python-dotenv>=1.0.1
faker>=24.0.0
requests>=2.31.0
playwright>=1.40.0
